requirements.txt
celery==4.4.7
Django==3.1.8
django-celery-beat==2.1.0
python-dateutil==2.8.1
redis==3.5.3
djangorestframework==3.11.2
django-rest-auth==1.10
django-rest-swagger
django-allauth